Reorganized the folders
There is now a Unity project folder, in case anyone wants to see my work, and a demo folder. The demo folder has an .EXE that you can run on windows, mac, or Linux. You must have two Microsoft, wired, Xbox 360 controllers (for some reason Unity only likes that kind of gamepad) to play. The game is also meant to play similar to the original Pong arcade game. The Monitor is laying on a table, and players are on opposite sides of the monitor. The monitor should be vertical to the player.